+++ NEWS.md
@@ -1,6 +1,7 @@
 # genderBR (development version)
 
 - Added support for IBGE's 2022 names API via a new `year` argument in `get_gender`.
+- Internal dataset `nomes` now includes probabilities for 2010 and 2022 (`prob_fem10`, `prob_fem22`) and is used when `internal = TRUE`.
 
 # genderBR 1.2.0